First attempt. Implemented in R using the 'raster' package.
The new R script biomass_model.R implements the basic model (linear combination
of input data stored in large geotiff files) using the 'raster' package to do
the heavy lifting. 'raster' is known not to be terribly efficient, but it
greatly simpifies this task by handling all the reading and writing behind the
scenes, as well as automatically breaking the data up into managable chunks.

To run on SCC, I first load the following modules: R/R-3.1.1, gdal, proj4. It
may or may not work in other environments