Fixed projection of output file
For some reason, the 'raster' package changes the crs when writing a file to
disk. I could not change this behavior, and I could not find a way to use rgdal to
rewrite the crs (without delving into undocumented subroutines). As an
annoyingly clunky workaround, I currently call an external python script
(provided as an example for the python GDAL bindings) to match the output and
input projections after the fact. This script is called from R to avoid adding
additional files.

Also, I added a shebang line to the script so that it can be called directly
from the command line, meaning either 'Rscript biomass_model.R' of
'./biomass_model.r' will do the trick.